+++ doc/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 Version 0.7.3
 -------------
 
+* convert: add ``--extract-metadata=NAME``
 * metadata: add ``mem_rss`` and ``mem_private`` metadat
 
 Version 0.7.2 (2016-07-21)